Initial commit for grib2io-interp package.
This package was originally a part of the grib2io package.  This
packages contains a Python interface to the NCEPLIBS-ip library.  Since
this library is Fortran-based, you must use NumPy's distutil's Extension
and setup to build and install.

This is led to too many complications with grib2io build/install process
which uses setuptools.