Another setup.py update
OK...I think we are good now.  In this update, in find_library, we
are getting the absolute path of the lib and resolving symlinks.

In the case of using libs from Homebrew, this will resolve to the
true install location in $(brew --prefix)/Cellar/

Also in find_library, added logic to check for when in a conda
installation via env var CONDA_PREFIX.